% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chumbley-non-random.R
\name{chumbley_non_random}
\alias{chumbley_non_random}
\title{Chumbley Non-Random}
\usage{
chumbley_non_random(data1, data2, window_opt = 500, window_val = 50,
  coarse = 0.25)
}
\arguments{
\item{data1}{The first tool mark as a 1-column matrix}

\item{data2}{The second tool mark as a 1-column matrix}

\item{window_opt}{size of the window to be used in the optimization step}

\item{window_val}{Size of the window to be used in the validation step}

\item{coarse}{smoothing parameter for the normalization smooth}
}
\value{
list with
\itemize{
\item {same_shift_n} Number of same shift offsets used
\item {diff_shift_n} Number of different shift offsets used
\item {U} observed U statistic
\item {p_value} Corresponding p-value
}
}
\description{
This function computes the Chumbley U-Statistic on systemically chosen pairs of windows rather 
than the original method which selects randomly chosen pairs of windows
}
\examples{
results <- chumbley_non_random(y_1, y_2, window_opt = 500, window_val = 50, coarse = .25)
}
